feat: Remove empty and duplicate columns
There are 369 feature columns in the raw data. Many of these are all zeros (or
at least almost all zeros) and others are duplicates. We remove them to speed
up the boosting classifier.

Now it should be noted that the method of removing these columns may be
overzealous. We remove any column that has a standard deviation of zero after
training data balancing , as well as any column that has the same mean and
standard deviation of a previous column. Removal is not consistent from run to run.